configgen: pass to configgen the number of axes of the devices
linux event handles axes and hats in only one way.
sdl (and probably the old joystick api) separates them in hats and axes.
if you have 2 axes and 1 hat,
dolphin joystick configuration file won't the id of the hat (1 in our case), but
the number of the axes among all axes (3 and 4 because a hat is 2 axes)
in the linux input.h config files, hats are at the end of the axes.
So, a "look like" normal way to compute the dolphin values is to give him the number
of axes (in the sdl sens of axes) of the joystick. It would have been saved in the es_input file,
but it can be computed automatically by es.

note that this break the buildroot compilation while it brakes a es patch.
It will be easier to fix once this commit will be applied.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>