exit on reboot required
on batocera side, i'll add a script to restart es in case of crash
to avoid the incidence of the black screen,
+ for some options, like language, video output, we can just restart es
so that it is taken immediatly. unfortunnatly, it is not perfectly consistent :
for example, i don't do it on the secure option or overclocking options while
a real reboot is required for them.

Signed-off-by: Nicolas Adenis-Lamarre <nicolas.adenis.lamarre@gmail.com>