Add resize option to decorations submenu + minor fixes
Allows the user to opt to use the original resolution + ratio of the image they have selected as the tattoo. "AUTO" behaviour is to resize to 225 px proportionately (similar to current live behaviour).
Also adds contingencies if the user tattoo is too large for the screen
and consistent terminology to some other items in the decorations submenu (so as not to confuse the term "overlay" (which is what RetroArch refers to bezels as sometimes) with "bezel")