Clarity and label improvements
This makes some more dramatic changes and adds in a lot of description labels into the interface. I am aiming to get some feedback on this before committing it to master.

The objective of this PR are as follows:
* To make the menus more friendly for new users who may not have experienced emulation software before.
* Reduce redundancy in labelling.
* Explain what the settings do without ambiguity.
* Reorganise the controller settings to show that there's more than just player order settings.

Since most strings are new, it would be fine to put this in for the next immediate release. However, some strings (most important, the "UI SETTINGS" string) have been changed and require retranslation to the new sting.

- All instances of UI to User Interface (where appropriate).
- More descriptions to confusing options.
- Incremental savestate option values relabelled and explained.