more accurate descriptive label for video mode
Complement to https://github.com/batocera-linux/batocera.linux/pull/6811

"The video mode description says 'Force the emulator to run at this resolution', but that's not entirely true. The emulator's rendering resolution is what decides your game's resolution, the video mode is just the resolution of the signal sent to the display.

The reason for this change was a new user thinking that video mode was the option to change their rendering resolution. This change aims to improve the new user experience for this setting."