update ping utility
Check for the HTTP URL of Google instead of only relying on its DNS. This will improve the robustness of the tool and also how quickly it resolves (in my personal testing at least, pinging the http is 2ms quicker).

Also made the structure of the code more consistent with the rest of the file.