Select running board as default installation target
A common source of user error when using the "INSTALL ON A NEW DISK"
feature is that the user guesses their architecture incorrectly.

With this change, if an installation architecture matches the
currently running board, it will be selected as default.

If no match is found, the previous behavior is retained.